requirements.txt
pyserial
mido
python-rtmidi
black
mypy
